Improved UI and added transition
I improved the UI of the menu screen by adding a grey background behind the menu with the alpha set low, in order to still see the city in the background, but not have it being too prominent. I also added the second transition (from TimeTravel_start to TimeTravel_end) and added the TimeTravel trigger to it. This will work once the TimeTravel branch ahs been merched into this branch.